[PATCH] Settings: Implement Active Display [2/2]
All credit to ChameleonOS

refer: http://forum.xda-developers.com/showthread.php?t=2450697

This is an implementation of the active display feature, similar
to that found on the Moto X.

When enabled the screen will turn on and display the
incoming notification.  The user can then open the app
for the notification, dismiss the notification or unlock
their device normally.

Additional notifications, if any exist, will be displayed
below the ring notification and can be viewed by touching
the notification icon.  The user can simply slide their
finger across these other notification to quickly view
them.  The last one touched will remain in the ring and
the user can then perform any of the actions listed above.

Change-Id: Ibafa1393eec3aeff9c1e6041dce9f127b4cc91b2